Add files via upload
This perl module can be used to generate simple page tops, and bottoms with simple strings, and can generate simple forms when 
passed an appropriate hash of lists. Examples of how this is done can be seen in the perl-fortran, perl-C++ repositories purple, enigma, and directform, and probably soon the avl tree web app too.